demo/sample-repo/requirements.txt
langchain==0.3.11
crewai==0.1.0
openai==1.54.3
anthropic==0.39.0
httpx==0.27.0
pyautogen==0.2.0
